[Chore #12629770] Clean up and refactor folder structure
