+++ app/static/components/forms/FolderForm.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Loader from 'react-loader';
-import * as action from './actions/GenericAction';
-import * as Toast from './actions/ToastAction';
-import store from './store/FolderStore';
+import * as action from '../actions/GenericAction';
+import * as Toast from '../actions/ToastAction';
+import store from '../store/FolderStore';
 
 export default class extends React.Component {
     constructor(){
