+++ app/static/components/pages/HomeLayout.jsx
@@ -3,16 +3,10 @@ import Nav from '../Nav';
 import Notifications from 'react-notify-toast';
 
 export default class extends React.Component {
-    constructor() {
-      super();
-      var index = $('#index').val();
-      this.state = {active: true ? index == 'Dashboard' : false}
-    }
-
     render() {
         return (
         <div class="row">
-        <Notifications />
+          <Notifications />
           <Nav user_active_class={'hide'} />
           <div class="container">
              {this.props.children}
