--- app/static/components/pages/Layout.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import Nav from '../Nav';
-
-export default class extends React.Component {
-    constructor() {
-      super();
-
-    }
-    componentDidMount(){
-
-    }
-
-    render() {
-        return (
-        <div class="row">
-
-          <Nav />
-          <div class="container">
-          <Notifications />
-             {this.props.children}
-          </div>
-        </div>
-        );
-    }
-}
