--- app/templates/test.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Photo Editing</title>
-    {% load staticfiles %}
-    <link rel="stylesheet" href="{% static 'css/themes/cosmo/bootstrap.min.css' %}" media="screen">
-    <link rel="stylesheet" href="{% static 'css/custom.min.css' %}">
-    <link href="{% static 'css/style4.css' %}" type="text/css" rel="stylesheet" />
-    <link href="{% static 'css/magnific-popup.css' %}" type="text/css" rel="stylesheet" />
-    <link href="{% static 'css/main.css' %}" type="text/css" rel="stylesheet" />
-</head>
-
-<body>
-<input type="hidden" name="index" id='index' value="{{index}}" />
-<input type="hidden" name="active-folder" id='active-folder' value="0" />
-<input type="hidden" name="currentUser" id='current-user' value="{{user}}" />
-<p>
-    hello
-</p>
-<p>
-    <img src="data:image/jpeg;base64, {{image}}" class="img-responsive" style="width:800px" />
-</p>
-
-<script type="text/javascript" src="{% static 'js/jquery-1.10.2.min.js' %}"></script>
-<script type="text/javascript" src="{% static 'js/bootstrap.min.js' %}"></script>
-<script type="text/javascript" src="{% static 'js/jquery.magnific-popup.min.js' %}"></script>
-</body>
-</html>
